backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
wikipedia==1.4.0
openai==0.27.0
python-dotenv==0.19.0
requests==2.26.0
